Do not use variable names ending on underscore for input variables and simpler warning suppression for generic_stringize example

+++ example/generic_stringize.cpp
@@ -55,7 +55,7 @@ int main() {
     boost::tuple<char, int, char, int> decim('-', 10, 'e', 5);
     assert(stringize(decim) == "-10e5");
 
-    std::pair<short, std::string> value_and_type(270, "Kelvin");
+    std::pair<int, std::string> value_and_type(270, "Kelvin");
     assert(stringize(value_and_type) == "270Kelvin");
 }
